docker-compose에서 celery 관련 오류 수정 및 celery 비동기 처리 구현.
